Add shade generation functionality to color spaces
This update introduces a shades generation feature in the HSL and RGB color spaces. With the implementation of `generateShades` method in these classes, users can now generate Material Design-inspired shades from any base color. The change includes necessary updates to the README file to guide users about this newly added feature. The colorist toolkit version is bumped to 1.0.5 to mark this change.