Refactor code to generate shades instead of custom palette
The process of creating custom palettes based on an LchColor has been refactored to generate shades instead. This more focused action reduces the complexity of the system and allows colors to be better represented. Changes include the modification of Colorist\Palettes\Shades, the changing from createPalette() to createShades(), the removal of generateFromBase() method, and the minor simplification of the existing createCustomPalette() function.